#194 rename InputAttributeMatchesParameter to InputAttributeHasMatchingParameter
