+++ CodeComplianceTest_Engine/Query/Checks/InputAttributeHasMatchingParameter.cs
@@ -20,22 +20,39 @@
  * along with this code. If not, see <https://www.gnu.org/licenses/lgpl-3.0.html>.      
  */
 
+using BH.oM.Test;
+using BH.oM.Test.Attributes;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
 using System;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
-
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace BH.Test.Test
+namespace BH.Engine.Test.CodeCompliance.Checks
 {
-    public partial class Test_Engine
+    public static partial class Query
     {
-        [TestMethod]
-        public void InputAttributeMatchesParameter()
+        [Message("Input attribute does not match any of the given parameters")]
+        [ErrorLevel(ErrorLevel.Error)]
+        [Path(@"([a-zA-Z0-9]+)_(Engine|Adapter)\\.*\.cs$")]
+        public static Span InputAttributeHasMatchingParameter(this AttributeSyntax node)
         {
-            Test.RunTest("InputAttributeMatchesParameter", GetChangedObjectFiles(), GetProjectName());
+            if (node.Name.ToString() != "Input") return null;
+
+            var method = node.Parent.Parent as BaseMethodDeclarationSyntax;
+            if (method != null && method.IsPublic() && (method.IsEngineMethod() || method.IsAdapterConstructor()))
+            {
+                if (node.ArgumentList.Arguments.Count >= 2)
+                {
+                    string paramname = node.ArgumentList.Arguments[0].Expression.GetFirstToken().Value.ToString();
+                    if (method.ParameterList.Parameters.Any((p) => p.Identifier.Text == paramname)) return null;
+                    else return node.ArgumentList.Arguments[0].Span.ToSpan();
+                }
+                return node.Span.ToSpan();
+            }
+            return null;
         }
     }
 }
+
